find weather by coordinates, checkGpsStatus, EasyPermissions

+++ app/src/main/java/com/skilled/weatherapp/repository/Repository.kt
@@ -4,5 +4,5 @@ import com.skilled.weatherapp.api.RetrofitInstance
 
 class Repository {
     suspend fun getWeatherByCityName(name: String) = RetrofitInstance.api.getWeatherByCityName(name = name)
-    suspend fun getWeatherByCoordinates(lat: Long, lon:Long) = RetrofitInstance.api.getWeatherByCoordinates(lat = lat, lon = lon)
+    suspend fun getWeatherByCoordinates(lat: Double, lon:Double) = RetrofitInstance.api.getWeatherByCoordinates(lat = lat, lon = lon)
 }
